+++ CHANGELOG.md
@@ -14,7 +14,7 @@ publication of Payroll Information as fname `earnings-statement-for-all`.
 + fix: open surplus earnings statements like opening the first ten earnings
   statements, not with JavaScript that'll trigger pop-up blockers.
 + style: obsessively 2-space-indent `payrollInformation.jsp` in the hope of
-  sooner detecting future bugs like HRSPLT-404.
+  sooner detecting future bugs like [HRSPLT-404][].
 
 ### 6.0.4 fix tax statements tab